Embedded scheduler working

+++ cmd/manager-migration/migrations/000001_create_schedule_table.up.sql
@@ -9,7 +9,7 @@ CREATE TABLE IF NOT EXISTS schedule
     network     VARCHAR(100)  NOT NULL,
     version     VARCHAR(50)  NOT NULL,
 
-    duration    INTEGER,
+    duration    BIGINT,
     kind        TEXT,
 
     PRIMARY KEY (id)
